Move some Node.js functions and add missing export
