+++ packages/utils/src/snaps.ts
@@ -1,4 +1,3 @@
-import { createHash } from 'crypto';
 import { SerializedEthereumRpcError } from 'eth-rpc-errors/dist/classes';
 import { Json } from '@metamask/utils';
 import { SnapManifest } from './json-schemas';
@@ -152,38 +151,7 @@ export class ProgrammaticallyFixableSnapError extends Error {
   }
 }
 
-/**
- * Calculates the Base64-encoded SHA-256 digest of a Snap source code string.
- *
- * @param sourceCode - The UTF-8 string source code of a Snap.
- * @returns The Base64-encoded SHA-256 digest of the source code.
- */
-export function getSnapSourceShasum(sourceCode: string): string {
-  return createHash('sha256').update(sourceCode, 'utf8').digest('base64');
-}
-
 export type ValidatedSnapId = `local:${string}` | `npm:${string}`;
-
-/**
- * Checks whether the `source.shasum` property of a Snap manifest matches the
- * shasum of a snap source code string.
- *
- * @param manifest - The manifest whose shasum to validate.
- * @param sourceCode - The source code of the snap.
- * @param errorMessage - The error message to throw if validation fails.
- */
-export function validateSnapShasum(
-  manifest: SnapManifest,
-  sourceCode: string,
-  errorMessage = 'Invalid Snap manifest: manifest shasum does not match computed shasum.',
-): void {
-  if (manifest.source.shasum !== getSnapSourceShasum(sourceCode)) {
-    throw new ProgrammaticallyFixableSnapError(
-      errorMessage,
-      SnapValidationFailureReason.ShasumMismatch,
-    );
-  }
-}
 
 /**
  * Extracts the snap prefix from a snap ID.
